requirements.txt
asgiref==3.6.0
autopep8==2.0.0
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
